# Warnings, docs, benchmarks and performance optimizations