--- a/prometheus-port.cabal
+++ b/prometheus-port.cabal
@@ -81,3 +81,23 @@ test-suite consistency
                , prometheus-port
 
   default-language: Haskell2010
+
+benchmark time
+  type: exitcode-stdio-1.0
+
+  main-is: Time.hs
+
+  hs-source-dirs: benchmarks
+
+  build-depends: base >=4.7 && <5
+               , bytestring
+               , deepseq
+               , gauge
+               , prometheus-port
+
+  ghc-options: -O2
+               -threaded
+               -rtsopts
+               -with-rtsopts=-N
+
+  default-language: Haskell2010