--- a/prometheus-port.cabal
+++ b/prometheus-port.cabal
@@ -94,6 +94,7 @@ benchmark time
                , deepseq
                , gauge
                , prometheus-port
+               , random
 
   ghc-options: -O2
                -threaded