--- a/prometheus-port.cabal
+++ b/prometheus-port.cabal
@@ -26,7 +26,8 @@ library
 
   hs-source-dirs: src
 
-  ghc-options:         -Wall
+  ghc-options:         -O2
+                       -Wall
                        -Wcompat
                        -Widentities
                        -Wincomplete-uni-patterns