--- a/prometheus-port.cabal
+++ b/prometheus-port.cabal
@@ -26,7 +26,7 @@ library
 
   hs-source-dirs: src
 
-  ghc-options:         -O2
+  ghc-options:         -O1
                        -Wall
                        -Wcompat
                        -Widentities