--- a/src/Network/Wai/Middleware/Prometheus.hs
+++ b/src/Network/Wai/Middleware/Prometheus.hs
@@ -31,7 +31,7 @@ newtype HttpMetrics f =
 {-# NOINLINE httpMetrics #-}
 httpMetrics :: HttpMetrics Metric
 httpMetrics =
-  let buckets = take 14 . iterate (*2) $ 1 / 1024
+  let buckets = take 14 . iterate (*2) $! 1 / 1024
   in HttpMetrics
        { hmDuration = vector ("path", "method", "status") $
                         histogram (Info "http_req_duration" "HTTP request duration (in seconds)") buckets