--- a/src/Prometheus/Internal/Primitive.hs
+++ b/src/Prometheus/Internal/Primitive.hs
@@ -29,6 +29,7 @@ import qualified Prometheus.Internal.Pure.Histogram as Pure (Bucket, Histogram)
 import qualified Prometheus.Internal.Pure.Summary as Pure (Quantile, Estimator, Summary)
 
 import           Control.Concurrent.STM.TVar
+import           Control.DeepSeq
 import           Control.Monad.STM
 import           Data.Functor.Identity
 import           Data.Proxy
@@ -38,25 +39,32 @@ register_ :: Rank a ~ Identity => Metric a -> IO (Impure () TVar (Purify a))
 register_ (Metric (Impure _def info base)) = do
   tvar <- newTVarIO $ runIdentity base
   return $ Impure () info tvar
+{-# INLINABLE register_ #-}
 
 extract_ :: Pure.Extract a e => Impure d TVar a -> IO e
 extract_ (Impure _ _ t) = Pure.extract <$> readTVarIO t
+{-# INLINABLE extract_ #-}
 
 export_ :: (Pure.Name a, Pure.Export a) => Impure d TVar a -> IO Template
 export_ (Impure _ info t :: Impure d TVar a) =
   Template info (Pure.name (Proxy :: Proxy a)) . Pure.export <$> readTVarIO t
+{-# INLINABLE export_ #-}
 
-plus_ :: (Pure.Increment a) => Double -> Impure d TVar a -> IO ()
-plus_ a (Impure _ _ t) = atomically . modifyTVar' t $ Pure.plus a
+plus_ :: (NFData a, Pure.Increment a) => Double -> Impure d TVar a -> IO ()
+plus_ a (Impure _ _ t) = atomically . modifyTVar' t $ force . Pure.plus a
+{-# INLINABLE plus_ #-}
 
-minus_ :: (Pure.Decrement a) => Double -> Impure d TVar a -> IO ()
-minus_ a (Impure _ _ t) = atomically . modifyTVar' t $ Pure.minus a
+minus_ :: (NFData a, Pure.Decrement a) => Double -> Impure d TVar a -> IO ()
+minus_ a (Impure _ _ t) = atomically . modifyTVar' t $ force . Pure.minus a
+{-# INLINABLE minus_ #-}
 
-set_ :: (Pure.Set a) => Double -> Impure d TVar a -> IO ()
-set_ a (Impure _ _ t) = atomically . modifyTVar' t $ Pure.set a
+set_ :: (NFData a, Pure.Set a) => Double -> Impure d TVar a -> IO ()
+set_ a (Impure _ _ t) = atomically . modifyTVar' t $ force . Pure.set a
+{-# INLINABLE set_ #-}
 
 observe_ :: (Pure.Observe a) => Double -> Impure d TVar a -> IO ()
 observe_ a (Impure _ _ t) = atomically . modifyTVar' t $ Pure.observe a
+{-# INLINABLE observe_ #-}
 
 
 
@@ -73,21 +81,26 @@ type instance Extra Counter = Identity
 
 instance Register Counter where
   register = fmap Counter . register_
+  {-# INLINEABLE register #-}
 
 instance Extract Counter Double where
   extract = extract_ . unCounter
+  {-# INLINEABLE extract #-}
 
 instance Export Counter where
   export = export_ . unCounter
+  {-# INLINEABLE export #-}
 
 instance Increment Counter where
   plus a = plus_ a . unCounter
+  {-# INLINEABLE plus #-}
 
 -- | The [doc](https://prometheus.io/docs/instrumenting/writing_clientlibs/) doesn't require
 --   this. It is useful for when the value is known to be monotonic, but you have no control
 --   over it, such as with GHC RTS values sampled by "Prometheus.GHC".
 instance Set Counter where
   set a (Counter (Impure _ _ t)) = atomically . modifyTVar' t $ Pure.set a
+  {-# INLINEABLE set #-}
 
 counter :: Info -> Metric Counter
 counter info = Metric $ let new = Pure.construct ()
@@ -106,21 +119,27 @@ type instance Extra Gauge = Identity
 
 instance Register Gauge where
   register = fmap Gauge . register_
+  {-# INLINEABLE register #-}
 
 instance Extract Gauge Double where
   extract = extract_ . unGauge
+  {-# INLINEABLE extract #-}
 
 instance Export Gauge where
   export = export_ . unGauge
+  {-# INLINEABLE export #-}
 
 instance Increment Gauge where
   plus a = plus_ a . unGauge
+  {-# INLINEABLE plus #-}
 
 instance Decrement Gauge where
   minus a = minus_ a . unGauge
+  {-# INLINEABLE minus #-}
 
 instance Set Gauge where
   set a = set_ a . unGauge
+  {-# INLINEABLE set #-}
 
 gauge :: Info -> Metric Gauge
 gauge info = Metric $ let new = Pure.construct ()
@@ -139,15 +158,19 @@ type instance Extra Histogram = Identity
 
 instance Register Histogram where
   register = fmap Histogram . register_
+  {-# INLINEABLE register #-}
 
 instance Extract Histogram Pure.Histogram where
   extract = extract_ . unHistogram
+  {-# INLINEABLE extract #-}
 
 instance Export Histogram where
   export = export_ . unHistogram
+  {-# INLINEABLE export #-}
 
 instance Observe Histogram where
   observe a = observe_ a . unHistogram
+  {-# INLINEABLE observe #-}
 
 histogram :: Info -> [Pure.Bucket] -> Metric Histogram
 histogram info buckets = Metric $ let new = Pure.construct buckets
@@ -166,15 +189,19 @@ type instance Extra Summary = Identity
 
 instance Register Summary where
   register = fmap Summary . register_
+  {-# INLINEABLE register #-}
 
 instance Extract Summary Pure.Summary where
   extract = extract_ . unSummary
+  {-# INLINEABLE extract #-}
 
 instance Export Summary where
   export = export_ . unSummary
+  {-# INLINEABLE export #-}
 
 instance Observe Summary where
   observe a = observe_ a . unSummary
+  {-# INLINEABLE observe #-}
 
 summary :: Info -> [Pure.Quantile] -> Metric Summary
 summary info quantiles = Metric $ let new = Pure.construct quantiles