Includes:
- Benchmarks on `Prometheus.Internal.Pure` and `Prometheus.Internal.Primitives`.
- Performance improvements on `Prometheus.Internal.Pure` and `Prometheus.Internal.Primitives`.
- Latests ghc warning flags; with corresponding fixes.
- More docs and haddocks.
- Improves `README.md`.

# Benchmarks

I am applying micro-optimizations i.e. not refactoring the implementation. Notice, all the benchmarks have noticeable better performance on the _after_  version.

`Estimator/Summary` is very quick on `extract/export`, but slow on`observe`. This has been achieved by materializing the final result on each `observe` i.e. the cost of `extract/export` has been amortized on each `observe`.

> I think there is room for improvement on `Pure.Summary.hs` by redoing the algorithm.

## Before

```
benchmarked Pure/Increment/Counter: 10000
time                 69.78 μs   (52.76 μs .. 81.18 μs)
                     0.791 R²   (0.678 R² .. 0.892 R²)
mean                 71.92 μs   (66.13 μs .. 80.11 μs)
std dev              24.85 μs   (19.29 μs .. 32.70 μs)
variance introduced by outliers: 95% (severely inflated)

benchmarked Pure/Increment/Counter: 100000
time                 567.2 μs   (468.4 μs .. 655.1 μs)
                     0.796 R²   (0.666 R² .. 0.873 R²)
mean                 789.7 μs   (694.1 μs .. 921.0 μs)
std dev              397.6 μs   (239.1 μs .. 571.4 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Increment/Gauge: 10000
time                 43.78 μs   (36.26 μs .. 53.80 μs)
                     0.653 R²   (0.519 R² .. 0.792 R²)
mean                 67.97 μs   (60.05 μs .. 76.57 μs)
std dev              29.25 μs   (24.49 μs .. 34.98 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Increment/Gauge: 100000
time                 435.7 μs   (406.8 μs .. 482.9 μs)
                     0.955 R²   (0.924 R² .. 0.997 R²)
mean                 428.1 μs   (418.7 μs .. 441.8 μs)
std dev              40.46 μs   (27.11 μs .. 58.92 μs)
variance introduced by outliers: 60% (severely inflated)

benchmarked Pure/Observe/Histogram: 10000
time                 1.019 ms   (922.3 μs .. 1.117 ms)
                     0.914 R²   (0.848 R² .. 0.969 R²)
mean                 1.159 ms   (1.092 ms .. 1.234 ms)
std dev              227.4 μs   (172.1 μs .. 285.7 μs)
variance introduced by outliers: 88% (severely inflated)

benchmarked Pure/Observe/Histogram: 100000
time                 35.82 ms   (32.54 ms .. 41.47 ms)
                     0.942 R²   (0.881 R² .. 0.985 R²)
mean                 35.21 ms   (33.56 ms .. 37.38 ms)
std dev              4.039 ms   (2.966 ms .. 5.404 ms)
variance introduced by outliers: 46% (moderately inflated)

benchmarked Pure/Observe/Summary: 10000
time                 46.94 ms   (37.57 ms .. 56.97 ms)
                     0.914 R²   (0.875 R² .. 0.995 R²)
mean                 42.16 ms   (38.75 ms .. 47.14 ms)
std dev              8.796 ms   (4.773 ms .. 13.85 ms)
variance introduced by outliers: 72% (severely inflated)

benchmarking Pure/Observe/Summary: 100000 ... took 28.87 s, total 56 iterations
benchmarked Pure/Observe/Summary: 100000
time                 601.3 ms   (458.1 ms .. 708.2 ms)
                     0.946 R²   (0.889 R² .. 0.988 R²)
mean                 500.9 ms   (453.4 ms .. 549.9 ms)
std dev              81.45 ms   (60.77 ms .. 109.7 ms)
variance introduced by outliers: 58% (severely inflated)

benchmarked Pure/Export/Counter: 10000
time                 8.825 ns   (7.918 ns .. 9.735 ns)
                     0.901 R²   (0.835 R² .. 0.947 R²)
mean                 11.99 ns   (10.56 ns .. 13.73 ns)
std dev              5.577 ns   (4.233 ns .. 7.207 ns)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Export/Counter: 100000
time                 13.40 ns   (9.941 ns .. 17.82 ns)
                     0.618 R²   (0.466 R² .. 0.787 R²)
mean                 13.27 ns   (11.94 ns .. 14.56 ns)
std dev              4.301 ns   (3.248 ns .. 5.540 ns)
variance introduced by outliers: 94% (severely inflated)

benchmarked Pure/Export/Histogram: 10000
time                 5.488 μs   (4.738 μs .. 6.360 μs)
                     0.822 R²   (0.754 R² .. 0.899 R²)
mean                 7.470 μs   (6.497 μs .. 9.450 μs)
std dev              4.191 μs   (3.113 μs .. 5.904 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Export/Histogram: 100000
time                 5.443 μs   (4.342 μs .. 6.599 μs)
                     0.830 R²   (0.758 R² .. 0.916 R²)
mean                 5.982 μs   (5.463 μs .. 6.872 μs)
std dev              2.335 μs   (1.756 μs .. 3.258 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Export/Summary: 10000
time                 9.318 μs   (6.811 μs .. 11.49 μs)
                     0.708 R²   (0.535 R² .. 0.866 R²)
mean                 8.216 μs   (7.318 μs .. 9.692 μs)
std dev              3.386 μs   (2.201 μs .. 5.314 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Pure/Export/Summary: 100000
time                 10.72 μs   (9.144 μs .. 12.98 μs)
                     0.796 R²   (0.724 R² .. 0.863 R²)
mean                 8.799 μs   (7.780 μs .. 9.858 μs)
std dev              3.473 μs   (3.056 μs .. 4.234 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Impure/Increment/Counter: 10000
time                 297.1 μs   (267.6 μs .. 338.5 μs)
                     0.869 R²   (0.822 R² .. 0.915 R²)
mean                 327.0 μs   (295.6 μs .. 367.7 μs)
std dev              113.3 μs   (82.60 μs .. 157.5 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Impure/Increment/Counter: 100000
time                 2.872 ms   (2.462 ms .. 3.247 ms)
                     0.892 R²   (0.831 R² .. 0.947 R²)
mean                 2.887 ms   (2.718 ms .. 3.085 ms)
std dev              566.8 μs   (462.7 μs .. 735.4 μs)
variance introduced by outliers: 87% (severely inflated)

benchmarked Impure/Increment/Gauge: 10000
time                 289.1 μs   (226.2 μs .. 353.7 μs)
                     0.794 R²   (0.691 R² .. 0.886 R²)
mean                 308.1 μs   (273.7 μs .. 384.9 μs)
std dev              156.4 μs   (59.50 μs .. 300.6 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Impure/Increment/Gauge: 100000
time                 2.584 ms   (2.326 ms .. 2.836 ms)
                     0.856 R²   (0.749 R² .. 0.926 R²)
mean                 3.185 ms   (2.960 ms .. 3.566 ms)
std dev              886.3 μs   (682.3 μs .. 1.151 ms)
variance introduced by outliers: 92% (severely inflated)

benchmarked Impure/Observe/Histogram: 10000
time                 1.999 ms   (1.783 ms .. 2.248 ms)
                     0.804 R²   (0.656 R² .. 0.927 R²)
mean                 2.095 ms   (1.868 ms .. 2.471 ms)
std dev              963.1 μs   (623.9 μs .. 1.521 ms)
variance introduced by outliers: 97% (severely inflated)

benchmarked Impure/Observe/Histogram: 100000
time                 19.02 ms   (15.73 ms .. 22.59 ms)
                     0.849 R²   (0.751 R² .. 0.932 R²)
mean                 21.45 ms   (19.01 ms .. 24.16 ms)
std dev              7.091 ms   (5.382 ms .. 9.763 ms)
variance introduced by outliers: 90% (severely inflated)

benchmarked Impure/Observe/Summary: 10000
time                 49.30 ms   (38.62 ms .. 61.04 ms)
                     0.894 R²   (0.822 R² .. 0.982 R²)
mean                 49.35 ms   (45.54 ms .. 52.58 ms)
std dev              7.427 ms   (5.781 ms .. 9.885 ms)
variance introduced by outliers: 58% (severely inflated)

benchmarking Impure/Observe/Summary: 100000 ... took 29.88 s, total 56 iterations
benchmarked Impure/Observe/Summary: 100000
time                 518.5 ms   (493.8 ms .. 550.4 ms)
                     0.985 R²   (0.952 R² .. 0.998 R²)
mean                 557.1 ms   (531.1 ms .. 590.9 ms)
std dev              53.40 ms   (40.71 ms .. 77.79 ms)
variance introduced by outliers: 28% (moderately inflated)
```

## After

```
benchmarked Pure/Increment/Counter: 10000
time                 81.66 μs   (63.72 μs .. 105.5 μs)
                     0.782 R²   (0.723 R² .. 0.875 R²)
mean                 70.19 μs   (63.30 μs .. 79.49 μs)
std dev              23.85 μs   (18.33 μs .. 30.30 μs)
variance introduced by outliers: 95% (severely inflated)

benchmarked Pure/Increment/Counter: 100000
time                 540.0 μs   (461.8 μs .. 634.5 μs)
                     0.789 R²   (0.679 R² .. 0.883 R²)
mean                 799.7 μs   (729.0 μs .. 875.2 μs)
std dev              275.0 μs   (229.2 μs .. 348.4 μs)
variance introduced by outliers: 95% (severely inflated)

benchmarked Pure/Increment/Gauge: 10000
time                 52.99 μs   (43.08 μs .. 67.09 μs)
                     0.730 R²   (0.637 R² .. 0.841 R²)
mean                 70.56 μs   (63.91 μs .. 80.52 μs)
std dev              24.17 μs   (17.88 μs .. 33.66 μs)
variance introduced by outliers: 94% (severely inflated)

benchmarked Pure/Increment/Gauge: 100000
time                 512.9 μs   (426.3 μs .. 591.8 μs)
                     0.769 R²   (0.669 R² .. 0.900 R²)
mean                 783.6 μs   (692.7 μs .. 901.5 μs)
std dev              381.0 μs   (280.3 μs .. 486.7 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Observe/Histogram: 10000
time                 535.4 μs   (489.1 μs .. 595.8 μs)
                     0.931 R²   (0.889 R² .. 0.968 R²)
mean                 572.9 μs   (544.4 μs .. 618.4 μs)
std dev              130.7 μs   (105.6 μs .. 178.9 μs)
variance introduced by outliers: 91% (severely inflated)

benchmarked Pure/Observe/Histogram: 100000
time                 31.64 ms   (26.53 ms .. 37.15 ms)
                     0.878 R²   (0.755 R² .. 0.955 R²)
mean                 29.82 ms   (27.61 ms .. 32.12 ms)
std dev              5.455 ms   (4.324 ms .. 7.137 ms)
variance introduced by outliers: 69% (severely inflated)

benchmarked Pure/Observe/Summary: 10000
time                 43.55 ms   (31.65 ms .. 56.41 ms)
                     0.801 R²   (0.585 R² .. 0.947 R²)
mean                 49.54 ms   (43.87 ms .. 54.05 ms)
std dev              11.76 ms   (8.436 ms .. 16.52 ms)
variance introduced by outliers: 79% (severely inflated)

benchmarking Pure/Observe/Summary: 100000 ... took 22.38 s, total 56 iterations
benchmarked Pure/Observe/Summary: 100000
time                 404.0 ms   (138.0 ms .. 632.5 ms)
                     0.597 R²   (0.291 R² .. 0.971 R²)
mean                 430.1 ms   (346.0 ms .. 526.3 ms)
std dev              145.7 ms   (113.8 ms .. 180.4 ms)
variance introduced by outliers: 89% (severely inflated)

benchmarked Pure/Export/Counter: 10000
time                 20.96 ns   (14.92 ns .. 27.23 ns)
                     0.602 R²   (0.442 R² .. 0.747 R²)
mean                 28.07 ns   (24.70 ns .. 32.60 ns)
std dev              13.68 ns   (10.97 ns .. 18.24 ns)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Export/Counter: 100000
time                 16.32 ns   (11.14 ns .. 20.64 ns)
                     0.567 R²   (0.359 R² .. 0.740 R²)
mean                 30.30 ns   (26.28 ns .. 35.31 ns)
std dev              15.89 ns   (13.48 ns .. 19.17 ns)
variance introduced by outliers: 98% (severely inflated)

benchmarked Pure/Export/Histogram: 10000
time                 6.361 μs   (4.354 μs .. 8.277 μs)
                     0.556 R²   (0.363 R² .. 0.710 R²)
mean                 9.281 μs   (8.098 μs .. 10.82 μs)
std dev              4.586 μs   (3.733 μs .. 5.958 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Pure/Export/Histogram: 100000
time                 5.692 μs   (4.305 μs .. 6.700 μs)
                     0.728 R²   (0.566 R² .. 0.849 R²)
mean                 8.344 μs   (7.430 μs .. 9.632 μs)
std dev              3.726 μs   (2.828 μs .. 4.759 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Pure/Export/Summary: 10000
time                 10.26 μs   (7.105 μs .. 13.24 μs)
                     0.673 R²   (0.473 R² .. 0.832 R²)
mean                 14.27 μs   (12.29 μs .. 15.60 μs)
std dev              5.943 μs   (4.888 μs .. 6.912 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Pure/Export/Summary: 100000
time                 8.883 μs   (6.813 μs .. 11.30 μs)
                     0.664 R²   (0.524 R² .. 0.796 R²)
mean                 10.18 μs   (8.975 μs .. 11.81 μs)
std dev              4.232 μs   (3.591 μs .. 5.140 μs)
variance introduced by outliers: 97% (severely inflated)

benchmarked Impure/Increment/Counter: 10000
time                 248.3 μs   (220.9 μs .. 276.6 μs)
                     0.912 R²   (0.882 R² .. 0.941 R²)
mean                 324.1 μs   (298.6 μs .. 368.1 μs)
std dev              128.9 μs   (87.95 μs .. 173.8 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Impure/Increment/Counter: 100000
time                 2.704 ms   (2.461 ms .. 2.960 ms)
                     0.928 R²   (0.878 R² .. 0.969 R²)
mean                 2.722 ms   (2.588 ms .. 2.906 ms)
std dev              522.4 μs   (425.3 μs .. 749.9 μs)
variance introduced by outliers: 87% (severely inflated)

benchmarked Impure/Increment/Gauge: 10000
time                 246.6 μs   (217.4 μs .. 290.1 μs)
                     0.866 R²   (0.814 R² .. 0.917 R²)
mean                 336.6 μs   (305.0 μs .. 375.8 μs)
std dev              120.7 μs   (84.13 μs .. 154.8 μs)
variance introduced by outliers: 98% (severely inflated)

benchmarked Impure/Increment/Gauge: 100000
time                 2.802 ms   (2.578 ms .. 3.020 ms)
                     0.939 R²   (0.890 R² .. 0.969 R²)
mean                 2.732 ms   (2.624 ms .. 2.914 ms)
std dev              433.0 μs   (364.4 μs .. 540.4 μs)
variance introduced by outliers: 80% (severely inflated)

benchmarking Impure/Observe/Histogram: 10000 ... took 9.566 s, total 2390 iterations
benchmarked Impure/Observe/Histogram: 10000
time                 517.4 μs   (386.3 μs .. 649.5 μs)
                     0.903 R²   (0.828 R² .. 0.995 R²)
mean                 391.7 μs   (379.7 μs .. 411.6 μs)
std dev              27.42 μs   (14.07 μs .. 41.78 μs)
variance introduced by outliers: 19% (moderately inflated)

benchmarked Impure/Observe/Histogram: 100000
time                 3.954 ms   (3.301 ms .. 4.683 ms)
                     0.897 R²   (0.769 R² .. 0.986 R²)
mean                 4.223 ms   (4.069 ms .. 4.444 ms)
std dev              389.8 μs   (228.6 μs .. 516.6 μs)
variance introduced by outliers: 34% (moderately inflated)

benchmarked Impure/Observe/Summary: 10000
time                 62.32 ms   (48.20 ms .. 72.62 ms)
                     0.878 R²   (0.716 R² .. 0.965 R²)
mean                 54.42 ms   (46.04 ms .. 60.54 ms)
std dev              13.71 ms   (10.44 ms .. 19.86 ms)
variance introduced by outliers: 76% (severely inflated)

benchmarking Impure/Observe/Summary: 100000 ... took 26.30 s, total 56 iterations
benchmarked Impure/Observe/Summary: 100000
time                 504.7 ms   (399.2 ms .. 594.8 ms)
                     0.916 R²   (0.776 R² .. 0.990 R²)
mean                 464.3 ms   (399.2 ms .. 509.0 ms)
std dev              87.23 ms   (56.87 ms .. 126.6 ms)
variance introduced by outliers: 58% (severely inflated)
```