@monadplus 2022-09-05
> I do not know if most of the changes within this PR are needed.

None of them is __needed__. Nevertheless, all of them are good to have: proper warnings, more docs, benchmarks, small performance improvements.

> You've changed the code from directly applying force to a lot of bang patterns + inlining, but I'd argue all you need to do is prove that plus/minus/set/observe evaluate the underlying data structure fully before placing the TVar.

I have changed from a coarse-grained thunk evaluation to a fine-grained thunk evaluation. It is true that they are equivalent after forcing evaluation, but the intermediate thunks are still created on the coarse-grained implementation. Ultimately, the fine-grained implementation would always give the best performance. Furthermore, the coarse-grained implementation may suffer from trying to evaluate a thunk more than once.

Notice, that I also unpacked some fields and inlined (or mark as inlineable) most of the one liner or two liner functions. This is probably where the performance gains are.

> I do not know whether the benchmarks are trustworthy, as you might've missed the -O1 on the files

I haven't missed the `-O1`, notice the `-O2` in the benchmark section on `prometheus-port.cabal`. The benchmarks are trustworthy and can be executed on any machine. You are welcome to run the benchmarks with `cabal bench`.

>  I do not know what the impact on compilation times and executable size is with the abundance of INLINEs and flags added.

Compilation times are marginally slower (4.40s vs 4.70s) because now you need to generate more `.hi` files and the optimizer has more chances to optimize the code i.e. it takes longer because the generated code is faster. I would be surprised if the size of the binary isn't bigger, but the increase if probably marginal and I wouldn't bother about binary sizes when performance is 2x-4x faster in most benchmarks.