@BurningWitness 2022-09-05
> notice the -O2 in the benchmark section on prometheus-port.cabal

Doesn't this mean both versions are compiled with `-O2`? :thinking: