@monadplus 2022-09-05
New results (after using `-O2` on the library):
- `Increment`/`Pure.Increment`/ has equivalent performance.
- `Pure.Observe` is faster (probably due to unboxing and inlining).
- `Export` has equivalent performance.
- `Observe` is faster: 18ms vs 4ms (on average, I repeated the benchmark several times just in case).