@BurningWitness 2022-09-05
Can you retest the old version one more time with `-O1` on the library? If difference is marginal in that case too, I believe the only real issue that needs addressing here is the `Histogram` (the 4x speedup is massive, so this may be the leaky part).