@monadplus 2022-09-05
> Can you retest the old version one more time with -O1 on the library? 

Yes, I will post the results. I think that the gains are from inline, unpacking and using a strict map on `Histogram`. I think the changes using `BangPatterns` are not doing anything and making the code worse, I will revert those after benchmarking it.