@BurningWitness 2022-09-05
I'd say `INLINE`'s are mostly useless too, all you're doing is forcing the compiler to duplicate code instead of allowing it to figure it out on its own.

Also I don't know where precisely the speedup on Histograms is (both lazy and strict `Map`s utilize the same data structure, so `NFData` is the same instance across both) :thinking: