@monadplus 2022-09-05
> I'd say INLINEs are mostly useless too, all you're doing is forcing the compiler to duplicate code instead of allowing it to figure it out on its own.

That is not 100% true. You are not duplicating any code. Apart from that `INLINEABLE` is super important to trigger specialization. Recall that "By default, functions are not specialised across modules. "! (for more info read https://wiki.haskell.org/Inlining_and_Specialisation)