@BurningWitness 2022-09-05
> you're not duplicating any code

As per the ancient [user guide](https://downloads.haskell.org/~ghc/7.0.3/docs/html/users_guide/pragmas.html):
GHC (with -O, as always) tries to inline (or “unfold”) functions/values that are “small enough,” thus avoiding the call overhead and possibly exposing other more-wonderful optimisations. Normally, if GHC decides a function is “too expensive” to inline, it will not do so, nor will it export that unfolding for other modules to use.

I'm fine with `INLINEABLE`s.