@monadplus 2022-09-05
I am gonna remove all the bang patterns (except from Histogram) and `INLINE`s (leaving only `INLINEABLE`s).