@monadplus 2022-09-05
TL;DR

- `Histogram.observe` we gained 4x speed-up.
- `Summary.observe` we gained 1.5x speed-up.
- The rest has similar performance (there are no changes).

Here are the latest results from benchmarks (I have only posted the `Observe` since the rest has similar results):

```
benchmarked Impure/Observe/Histogram: 100000 (OLD)
time                 19.02 ms   (15.73 ms .. 22.59 ms)
                     0.849 R²   (0.751 R² .. 0.932 R²)
mean                 21.45 ms   (19.01 ms .. 24.16 ms)
std dev              7.091 ms   (5.382 ms .. 9.763 ms)
variance introduced by outliers: 90% (severely inflated)

benchmarked Impure/Observe/Histogram: 100000 (NEW)
time                 4.039 ms   (3.787 ms .. 4.324 ms)
                     0.980 R²   (0.962 R² .. 0.991 R²)
mean                 4.210 ms   (4.107 ms .. 4.351 ms)
std dev              235.5 μs   (157.8 μs .. 318.4 μs)
variance introduced by outliers: 19% (moderately inflated)

benchmarked Impure/Observe/Summary: 100000 (OLD)
time                 518.5 ms   (493.8 ms .. 550.4 ms)
                     0.985 R²   (0.952 R² .. 0.998 R²)
mean                 557.1 ms   (531.1 ms .. 590.9 ms)
std dev              53.40 ms   (40.71 ms .. 77.79 ms)
variance introduced by outliers: 28% (moderately inflated)

benchmarked Impure/Observe/Summary: 100000 (NEW)
time                 361.5 ms   (291.3 ms .. 398.8 ms)
                     0.963 R²   (0.862 R² .. 0.999 R²)
mean                 360.8 ms   (339.3 ms .. 375.6 ms)
std dev              29.06 ms   (15.06 ms .. 43.55 ms)
variance introduced by outliers: 19% (moderately inflated)
```