@BurningWitness 2022-09-05
Are you sure `-O2` is required on this version of the library then?